refactor: rename stores to store
