+++ cognite/neat/graph/loaders/_base.py
@@ -8,7 +8,7 @@ from cognite.client.data_classes.capabilities import Capability
 
 from cognite.neat.issues import IssueList, NeatIssue, NeatIssueList
 from cognite.neat.issues.errors import AuthorizationError
-from cognite.neat.stores import NeatGraphStore
+from cognite.neat.store import NeatGraphStore
 from cognite.neat.utils.auxiliary import class_html_doc
 from cognite.neat.utils.upload import UploadResult, UploadResultList
