+++ cognite/neat/store/_base.py
