+++ tests/tests_unit/graph/test_extractors/test_asset_extractor.py
@@ -3,7 +3,7 @@ from cognite.client.testing import monkeypatch_cognite_client
 
 from cognite.neat.constants import DEFAULT_NAMESPACE
 from cognite.neat.graph.extractors import AssetsExtractor
-from cognite.neat.stores import NeatGraphStore
+from cognite.neat.store import NeatGraphStore
 from cognite.neat.utils.auxiliary import create_sha256_hash
 from tests.config import CLASSIC_CDF_EXTRACTOR_DATA
