+++ tests/tests_unit/graph/test_loaders/test_dms_loader.py
@@ -1,7 +1,7 @@
 from cognite.neat.graph.extractors import AssetsExtractor
 from cognite.neat.graph.loaders import DMSLoader
 from cognite.neat.rules.importers import InferenceImporter
-from cognite.neat.stores import NeatGraphStore
+from cognite.neat.store import NeatGraphStore
 from tests.config import CLASSIC_CDF_EXTRACTOR_DATA
