+++ tests/tests_unit/rules/test_importers/test_inference_importer.py
@@ -3,7 +3,7 @@ from cognite.neat.graph.extractors import AssetsExtractor, RdfFileExtractor
 from cognite.neat.rules.importers import InferenceImporter
 from cognite.neat.rules.models.data_types import Json
 from cognite.neat.rules.models.entities import MultiValueTypeInfo
-from cognite.neat.stores import NeatGraphStore
+from cognite.neat.store import NeatGraphStore
 from tests.config import CLASSIC_CDF_EXTRACTOR_DATA
