+++ tests/tests_unit/test_graph/test_loaders/test_rdf2dms.py
@@ -2,7 +2,7 @@ import pytest
 
 from cognite.neat.graph.loaders import DMSLoader
 from cognite.neat.rules.importers import InferenceImporter
-from cognite.neat.stores import NeatGraphStore
+from cognite.neat.store import NeatGraphStore
 from tests.data import car
